requirements.txt
fredapi
numpy
plotly
pre-commit
pytest
python-dotenv
ruff
scipy
streamlit
yfinance